Add performOperation fn, update OPERATORS.foreach
performOperation function allows to do a single
operation which includes getting the answer,
updating mainscreen, and clearing the second
number for next operation.

Update OPERATORS.forEach so that users can do
an operation when 2 numbers are present and an
operator is clicked.

Update equal click event listener so it only
operates when two number are present by calling
performOperation function.